Rewrite the version updater modules to typescript
